Close probe socket after confirming connection
I misunderstood what `agent.createConnection()` did---I thought it created
a socket *in the connection pool*, ready for use by the next request.
It turns out that it just creates a socket and hands it off to the caller;
as a result, webd was holding on to a dangling TCP connection. For single-
threaded backends, this meant that the backend was busy trying to service
that connection and was therefore unable to handle the actual requests,
which were coming in by a different connection.

Since there doesn't seem to be a way to easily put a socket into the
agent's connection pool, just close it after the connection is established.
This creates an extraneous connection to the backend, but if the backend is
not robust against this it will also fall over at the first port-scan it
sees, so hopefully this will not cause any problems.